Improve model formatting, correct product box links
Made the photo and title of each product box a link rather than the whole
box, to prevent relocating to the product page on clicking the add to
basket button. Improved the model formatting by allowing the photos to
expand to the full size of their parent container, and increased the
short stories font size to increase readability. Also added a simple
footer for style and general practice.